Verneplikt should always use current Grunnbeløp
The calcuation for grunnlag based on Verneplikt does not use beregningsdato to determine which Grunnbeløp to use. It should always use the most currently available Grunnbeløp, which should be todays date.

This implementation seems quite messy. It feels significantly more intuitive that the verneplikt-calculation should be its own service, but then we also need to introduce another service for coordinating the two calculation services.

It is blocking Arena from completing their value chain tests of G-justering, so I'm pushing this through. I hope you can forgive me in the future :(